feat(back): store rating average in f64 for better JSON output
The output is now 1.0 MB instead of 1.2 MB.

+++ back/src/main.rs
@@ -14,7 +14,7 @@ fn read(path: &str) -> String {
 
 #[derive(Debug)]
 struct Rating {
-    average: f32,
+    average: f64,
     count: u32,
 }
